Start Screen w/ Buttons
Added a screen on startup with title and button clicking functionality. Easy, medium, and hard button has event listener and we can add the sudoku board creations in there.

Fonts and colors are subject to change. I just used defaults for now. You can change it in the pygame.font.Font